split out Phases::FetchClusterConfiguration
