+++ lib/pharos/phases/configure_client.rb
@@ -8,15 +8,6 @@ module Pharos
       REMOTE_FILE = "/etc/kubernetes/admin.conf"
 
       def call
-        save_config_locally
-        config_map = previous_config_map
-        return unless config_map
-
-        cluster_context['previous-config-map'] = config_map
-        cluster_context['previous-config'] = build_config(config_map)
-      end
-
-      def save_config_locally
         Dir.mkdir(config_dir, 0o700) unless Dir.exist?(config_dir)
 
         logger.info { "Fetching kubectl config ..." }
@@ -36,23 +27,6 @@ module Pharos
 
       def config_file
         Pharos::Kube.config_path(@config.api_endpoint)
-      end
-
-      # @param configmap [Kubeclient::Resource]
-      # @return [Pharos::Config]
-      def build_config(configmap)
-        cluster_config = Pharos::YamlFile.new(StringIO.new(configmap.data['cluster.yml']), override_filename: "#{@host}:cluster.yml").load
-        cluster_config['hosts'] ||= []
-        data = Pharos::ConfigSchema.build.call(cluster_config)
-        Pharos::Config.new(data)
-      end
-
-      # @return [Kubeclient::Resource]
-      def previous_config_map
-        kube_client = Pharos::Kube.client(@host.api_address)
-        kube_client.get_config_map('pharos-config', 'kube-system')
-      rescue Kubeclient::ResourceNotFoundError
-        nil
       end
     end
   end
